requirements.txt
aiohttp==3.11.10
beautifulsoup4==4.12.3
customtkinter==5.2.2
Pillow==11.0.0
playwright==1.49.1
